services/local-model/app/main.py
"""
Local AI Model Server
FastAPI-based inference server for dental image analysis
"""

import os
import time
import logging
from typing import Optional
from contextlib import asynccontextmanager

import numpy as np
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from PIL import Image
import io

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
)
logger = logging.getLogger(__name__)

# Configuration
MODEL_VERSION = os.getenv("MODEL_VERSION", "v1.0")
CONFIDENCE_THRESHOLD = float(os.getenv("CONFIDENCE_THRESHOLD", "0.85"))
MAX_IMAGE_SIZE = int(os.getenv("MAX_IMAGE_SIZE", "2048"))


class InferenceResult(BaseModel):
    """Standardized inference result"""
    version: str
    measurements: dict
    landmarks: list
    summary: str
    confidence: float


class HealthResponse(BaseModel):
    """Health check response"""
    status: str
    version: str
    model_loaded: bool


# Mock model class - replace with actual model loading
class DentalAnalysisModel:
    """
    Dental Image Analysis Model
    
    This is a placeholder implementation.
    Replace with actual model loading and inference code.
    """
    
    def __init__(self):
        self.loaded = False
        self.version = MODEL_VERSION
        
    def load(self):
        """Load the model"""
        logger.info(f"Loading model version {self.version}")
        # TODO: Load actual model weights
        self.loaded = True
        logger.info("Model loaded successfully")
        
    def predict(self, image: np.ndarray) -> dict:
        """
        Run inference on image
        
        Args:
            image: numpy array of the image
            
        Returns:
            Dictionary with analysis results
        """
        if not self.loaded:
            raise RuntimeError("Model not loaded")
        
        # TODO: Replace with actual model inference
        # This is mock data for demonstration
        return {
            "measurements": {
                "sna_angle": 82.5,
                "snb_angle": 79.2,
                "anb_angle": 3.3,
                "mandibular_plane_angle": 26.8,
                "upper_incisor_to_na": 4.2,
                "lower_incisor_to_nb": 4.8,
            },
            "landmarks": [
                {"x": 150, "y": 200, "label": "sella"},
                {"x": 180, "y": 250, "label": "nasion"},
                {"x": 200, "y": 300, "label": "a_point"},
                {"x": 220, "y": 350, "label": "b_point"},
                {"x": 250, "y": 400, "label": "pogonion"},
                {"x": 180, "y": 320, "label": "upper_incisor_tip"},
                {"x": 200, "y": 380, "label": "lower_incisor_tip"},
            ],
            "summary": "Skeletal Class I pattern with normal vertical proportions.",
            "confidence": 0.92,
        }


# Global model instance
model: Optional[DentalAnalysisModel] = None


@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan handler"""
    global model
    
    # Startup
    logger.info("Starting up Local Model Server...")
    model = DentalAnalysisModel()
    model.load()
    logger.info(f"Server ready - Model version: {MODEL_VERSION}")
    
    yield
    
    # Shutdown
    logger.info("Shutting down...")


# Create FastAPI app
app = FastAPI(
    title="Orca Dental AI - Local Model Server",
    description="Local inference server for dental image analysis",
    version=MODEL_VERSION,
    lifespan=lifespan,
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


def preprocess_image(image_bytes: bytes) -> np.ndarray:
    """
    Preprocess image for inference
    
    Args:
        image_bytes: Raw image bytes
        
    Returns:
        Preprocessed numpy array
    """
    try:
        # Open image
        image = Image.open(io.BytesIO(image_bytes))
        
        # Convert to RGB if necessary
        if image.mode != "RGB":
            image = image.convert("RGB")
        
        # Resize if too large
        width, height = image.size
        max_dim = max(width, height)
        if max_dim > MAX_IMAGE_SIZE:
            scale = MAX_IMAGE_SIZE / max_dim
            new_width = int(width * scale)
            new_height = int(height * scale)
            image = image.resize((new_width, new_height), Image.Resampling.LANCZOS)
            logger.info(f"Resized image from {width}x{height} to {new_width}x{new_height}")
        
        # Convert to numpy array
        img_array = np.array(image)
        
        return img_array
        
    except Exception as e:
        logger.error(f"Image preprocessing error: {e}")
        raise HTTPException(status_code=400, detail=f"Invalid image: {str(e)}")


@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Health check endpoint"""
    return HealthResponse(
        status="healthy" if model and model.loaded else "unhealthy",
        version=MODEL_VERSION,
        model_loaded=model.loaded if model else False,
    )


@app.post("/infer", response_model=InferenceResult)
async def infer(file: UploadFile = File(...)):
    """
    Run inference on uploaded image
    
    Args:
        file: Image file to analyze
        
    Returns:
        Analysis results
    """
    start_time = time.time()
    
    # Validate file
    if not file.content_type or not file.content_type.startswith("image/"):
        raise HTTPException(status_code=400, detail="File must be an image")
    
    try:
        # Read image
        image_bytes = await file.read()
        
        if len(image_bytes) == 0:
            raise HTTPException(status_code=400, detail="Empty file")
        
        # Preprocess
        img_array = preprocess_image(image_bytes)
        
        # Run inference
        result = model.predict(img_array)
        
        # Check confidence
        if result["confidence"] < CONFIDENCE_THRESHOLD:
            logger.warning(f"Low confidence: {result['confidence']}")
        
        processing_time = (time.time() - start_time) * 1000
        logger.info(f"Inference completed in {processing_time:.2f}ms")
        
        return InferenceResult(
            version=MODEL_VERSION,
            measurements=result["measurements"],
            landmarks=result["landmarks"],
            summary=result["summary"],
            confidence=result["confidence"],
        )
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Inference error: {e}")
        raise HTTPException(status_code=500, detail=f"Inference failed: {str(e)}")


@app.get("/")
async def root():
    """Root endpoint"""
    return {
        "service": "Orca Dental AI - Local Model Server",
        "version": MODEL_VERSION,
        "status": "running" if model and model.loaded else "not ready",
    }


if __name__ == "__main__":
    import uvicorn
    
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", "8000"))
    
    uvicorn.run(app, host=host, port=port)
